+++ ShipMonitor/ShipMonitor.cs
@@ -2021,5 +2021,26 @@ namespace EddiShipMonitor
             }
         }
     }
+
+    public class ShipRoleConverter : IValueConverter
+    {
+        public object Convert ( object value, Type targetType, object parameter, CultureInfo culture )
+        {
+            if ( value is Role role )
+            {
+                return role;
+            }
+            return null;
+        }
+
+        public object ConvertBack ( object value, Type targetType, object parameter, CultureInfo culture )
+        {
+            if ( value is Role role )
+            {
+                return role;
+            }
+            return Role.MultiPurpose;
+        }
+    }
 }